Cleaning up warnings re: h5py dtype

+++ tests/uvdata/test_uvh5.py
@@ -154,7 +154,7 @@ def test_read_miriad_write_uvh5_read_uvh5(paper_miriad, tmp_path, cat_type):
 
     # create the file so the clobber gets tested
     with h5py.File(testfile, "w") as h5file:
-        h5file.create_dataset("Test", list(range(10)))
+        h5file.create_dataset("Test", list(range(10)), dtype="f4")
 
     uv_in.write_uvh5(testfile, clobber=True)
     uv_out.read(testfile)
